ADD and Norm : init_exit.c -> init.c exit.c
